# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CCD School Music Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List all learning modules
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a learning module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a module by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "404":
          description: Not found
...
  /progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: Get all user progress records
      responses:
        "200":
          description: Progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Progress"
    post:
      operationId: upsertProgress
      tags: [progress]
      summary: Create or update progress for a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
  /progress/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [progress]
      summary: Mark a lesson as complete
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Progress"
...
components:
  schemas:
...
    Module:
      type: object
      required: [id, courseId, title, orderIndex, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        createdAt:
          type: string
    ModuleInput:
      type: object
      required: [courseId, title, orderIndex]
      properties:
        courseId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        orderIndex:
          type: integer
...
    Progress:
      type: object
      required: [id, lessonId, completed, updatedAt]
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        completed:
          type: boolean
        watchedSeconds:
          type: ["integer", "null"]
        quizScore:
          type: ["integer", "null"]
        updatedAt:
          type: string
    ProgressInput:
      type: object
      required: [lessonId]
      properties:
        lessonId:
          type: integer
        completed:
          type: boolean
        watchedSeconds:
          type: integer
        quizScore:
          type: integer
...